Refactored to extract Wasm "EventRegistration" to a new file
